requirements.txt
django==5.0.3
mysqlclient==2.2.4
python-decouple==3.8
djangorestframework==3.15.1